+++ Proyecto1DPOO/src/dataBase/Factura.java
@@ -6,9 +6,9 @@ import java.util.List;
 public class Factura {
 
 	private LocalDateTime fecha;
-	private float subtotal;
-	private float impuestos;
-	private float propina;
+	private double subtotal;
+	private double impuestos;
+	private double propina;
 	private List<DetalleVenta> detallesDeLaVenta;
 	
 }
